Improve the Dockerfile (#26)
* Dockerfile: Inline `APP_NAME`

Since `merge-gatekeeper` is necessarily hard-coded in the entrypoint, there's
little value in extracting it to an environment variable; if we changed it, the
entrypoint wouldn't exist. This commit therefore inlines it.

* Dockerfile: Declare env vars as args where possible

We don't need `ORG` and `REPO` in the container environment. This commit
therefore redefines them as build arguments. This has the side benefit of being
able to change them at build time, which could be useful for people using forks
of this repo.

* Dockerfile: Combine env var definitions

In order to reduce the number of image layers, this commit combines some
environment variable definitions into a single step.